Сделано окно и модель представления изменения данных учебной групы, добавлен метод в репозиторий обновления данных в бд.

+++ EducationERP/ViewModels/Modules/DeanRoom/DeanRoomViewModel.cs
@@ -22,8 +22,9 @@ namespace EducationERP.ViewModels.Modules.DeanRoom
         }
 
         public RaketaTCommand<FacultyVM> OpenWindowAddEducationGroupCommand { get; }
+        public RaketaTCommand<EducationGroupVM> OpenWindowChangeEducationGroupCommand { get; }
         public RaketaTCommand<EducationGroupVM> DeleteEducationGroupCommand { get; }
-        public RaketaCommand ExitCommand { get; set; }
+        public RaketaCommand ExitCommand { get; }
 
         IServiceView _serviceView;
         ITabControl _tabControl;
@@ -38,14 +39,23 @@ namespace EducationERP.ViewModels.Modules.DeanRoom
             GetFaculty(id);
 
             OpenWindowAddEducationGroupCommand = RaketaTCommand<FacultyVM>.Launch(OpenWindowAddEducationGroup);
+            OpenWindowChangeEducationGroupCommand = RaketaTCommand<EducationGroupVM>.Launch(OpenWindowChangeEducationGroup);
             DeleteEducationGroupCommand = RaketaTCommand<EducationGroupVM>.Launch(DeleteEducationGroup);
             ExitCommand = RaketaCommand.Launch(CloseTab);
         }
 
         void OpenWindowAddEducationGroup(FacultyVM facultyVM) =>
             _serviceView.Window<AddEducationGroupViewModel>(null, facultyVM).Modal();
+        void OpenWindowChangeEducationGroup(EducationGroupVM educationGroupVM)
+        {
+            if (educationGroupVM == null) return;
+            _serviceView.Window<ChangeEducationGroupViewModel>(null, educationGroupVM).Modal();
+        }
+
         async void DeleteEducationGroup(EducationGroupVM educationGroupVM)
         {
+            if (educationGroupVM == null) return;
+
             bool isDeleted = await _facultyRepository.DeleteEducationGroup(educationGroupVM.Id);
             if (isDeleted) FacultyVM.EducationGroups.Remove(educationGroupVM);
         }
@@ -67,7 +77,22 @@ namespace EducationERP.ViewModels.Modules.DeanRoom
                 var educationGroupVM = new EducationGroupVM
                 {
                     Id = educationGroup.Id,
-                    NameEducationGroup = educationGroup.NameEducationGroup
+                    CodeEducationGroup = educationGroup.CodeEducationGroup,
+                    NameEducationGroup = educationGroup.NameEducationGroup,
+                    LevelGroup = educationGroup.LevelGroup,
+                    FormGroup = educationGroup.FormGroup,
+                    TypeGroup = educationGroup.TypeGroup,
+                    Course = educationGroup.Course,
+                    MaxNumberStudents = educationGroup.MaxNumberStudents,
+                    CodeDirection = educationGroup.CodeDirection,
+                    NameDirection = educationGroup.NameDirection,
+                    CodeProfile = educationGroup.CodeProfile,
+                    NameProfile = educationGroup.NameProfile,
+                    NameCuratorGroup = educationGroup.NameCuratorGroup,
+                    NameHeadmanGroup = educationGroup.NameHeadmanGroup,
+                    Formed = educationGroup.Formed,
+                    DateOfFormed = educationGroup.DateOfFormed,
+                    Students = new()
                 };
                 FacultyVM.EducationGroups.Add(educationGroupVM);
             }
